Update exercise_sound.py
I changed the entire way audio is played. The song I chose was the classic Mario Death Song, and I approximated the notes to the best of my ability. This still needs to be workshopped. The notes operate in a (Frequency, Duration) syntax. The loop at the end reads the entire mario array, pauses when necessary by utilizing the quiet() function and plays notes accordingly.